// src/apis/models/FileInfo.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class FileInfo {
  's3Key': string;
  'bucket'?: string;
  'filename': string;
  'size': number;
  'downloadLink'?: string;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 's3Key',
      baseName: 's3Key',
      type: 'string',
      format: '',
    },
    {
      name: 'bucket',
      baseName: 'bucket',
      type: 'string',
      format: '',
    },
    {
      name: 'filename',
      baseName: 'filename',
      type: 'string',
      format: '',
    },
    {
      name: 'size',
      baseName: 'size',
      type: 'number',
      format: '',
    },
    {
      name: 'downloadLink',
      baseName: 'downloadLink',
      type: 'string',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return FileInfo.attributeTypeMap;
  }

  public constructor() {}
}